+++ src/cli/input_notes.rs
@@ -383,8 +383,8 @@ mod tests {
             rng,
             store,
             executor_store,
-        )
-        .unwrap();
+            true,
+        );
 
         // generate test data
         let assembler = TransactionKernel::assembler();
@@ -434,8 +434,8 @@ mod tests {
             rng,
             store,
             executor_store,
-        )
-        .unwrap();
+            true,
+        );
 
         import_note(&mut client, filename_path).unwrap();
         let imported_note_record: InputNoteRecord =
@@ -468,8 +468,8 @@ mod tests {
             rng,
             store,
             executor_store,
-        )
-        .unwrap();
+            true,
+        );
 
         // Ensure we get an error if no note is found
         let non_existent_note_id = "0x123456";